requirements.txt
Django==3.1.2
django-phone-field==1.8.1
djangorestframework==3.12.1
ipdb
django-cors-headers
